Demo: remove profiling instruction
Command `set profile=1;` is used to populate the output of `SHOW PROFILE;`.
For postgres, users need to enable timing in order to show the duration of
individual commands on command line. MySQL display timing by default.

Fix the output of `Battlefield` movie query to be consistent with other
MySQL outputs.